requirements.txt
Cython==0.26.1
h5py==2.10.0
matplotlib==3.0.3
numpy==1.19.5
opencv-python==4.1.2.30
pandas==0.24.2
Pillow==7.2.0
protobuf==3.13.0
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
toolz==0.11.1
torch==1.6.0
torchfile==0.1.0
torchvision==0.7.0
IPython==7.16.1
tqdm==4.61.1
